rename file and add some test case
