+++ mysql/util.go
@@ -13,6 +13,7 @@ import (
 	mrand "math/rand"
 	"runtime"
 	"strings"
+	"time"
 
 	"github.com/Masterminds/semver"
 	"github.com/go-mysql-org/go-mysql/utils"
@@ -132,7 +133,7 @@ func RandomBuf(size int) []byte {
 	// When this project supports golang 1.20 as a minimum, then this mrand.New(...)
 	// line can be eliminated and the random number can be generated by simply
 	// calling mrand.Intn()
-	random := mrand.New(mrand.NewSource(utils.Now().UTC().UnixNano()))
+	random := mrand.New(mrand.NewSource(time.Now().UTC().UnixNano()))
 	min, max := 30, 127
 	for i := 0; i < size; i++ {
 		buf[i] = byte(min + random.Intn(max-min))
