+++ utils/now.go
@@ -1,19 +1,7 @@
-//go:build linux || darwin
+//go:build !unix
 
 package utils
 
-import (
-	"syscall"
-	"time"
-)
+import "time"
 
-// Now is a faster method to get current time
-func Now() time.Time {
-	var tv syscall.Timeval
-	if err := syscall.Gettimeofday(&tv); nil != err {
-		// If it failed at syscall, use time package instead
-		return time.Now()
-	}
-
-	return time.Unix(0, syscall.TimevalToNsec(tv))
-}
+var Now = time.Now
