--- utils/now2.go
@@ -1,7 +1,0 @@
-//go:build !linux && !darwin
-
-package utils
-
-import "time"
-
-var Now = time.Now
