+++ utils/now_unix.go
@@ -1,0 +1,19 @@
+//go:build unix
+
+package utils
+
+import (
+	"syscall"
+	"time"
+)
+
+// Now is a faster method to get current time
+func Now() time.Time {
+	var tv syscall.Timeval
+	if err := syscall.Gettimeofday(&tv); nil != err {
+		// If it failed at syscall, use time package instead
+		return time.Now()
+	}
+
+	return time.Unix(0, syscall.TimevalToNsec(tv))
+}
